moved routes into dashboard.js
