+++ router/index.js
@@ -4,7 +4,7 @@ export const router = express.Router();
 // Controllers
 import { Login, submitLogin, Logout } from "../controllers/login.js";
 import { Register, submitRegister } from "../controllers/register.js";
-import { Dashboard, searchDashboard } from "../controllers/dashboard.js";
+import { Dashboard, searchDashboard, Start, Stop, Pause, Restart, Logs, Modal, Stats, Hide, Reset, Chart } from "../controllers/dashboard.js";
 import { Apps, appSearch } from "../controllers/apps.js";
 import { Users } from "../controllers/users.js";
 import { Images, removeImage } from "../controllers/images.js";
@@ -13,7 +13,7 @@ import { Volumes, removeVolume } from "../controllers/volumes.js";
 import { Account } from "../controllers/account.js";
 import { Variables } from "../controllers/variables.js";
 import { Settings } from "../controllers/settings.js";
-import { Supporters } from "../controllers/supporters.js";
+import { Supporters, Thanks } from "../controllers/supporters.js";
 import { Syslogs } from "../controllers/syslogs.js";
 import { Portal } from "../controllers/portal.js"
 
@@ -29,13 +29,22 @@ const auth = (req, res, next) => {
 // Routes
 router.get("/login", Login);
 router.post("/login", submitLogin);
-router.get("/logout", Logout);
-
 router.get("/register", Register);
 router.post("/register", submitRegister);  
+router.get("/logout", Logout);
 
 router.get("/", auth, Dashboard);
-router.post("/", auth, searchDashboard);
+router.post("/start", auth, Start);
+router.post("/stop", auth, Stop);
+router.post("/pause", auth, Pause);
+router.post("/restart", auth, Restart);
+router.get("/logs", auth, Logs);
+router.get ("/modal", auth, Modal);
+router.get("/stats", auth, Stats);
+router.post("/hide", auth, Hide);
+router.post("/reset", auth, Reset);
+router.get("/chart", auth, Chart);
+
 
 router.get("/images", auth, Images);
 router.post("/removeImage", removeImage);
@@ -58,7 +67,9 @@ router.get("/syslogs", auth, Syslogs);
 router.get("/account", Account);
 router.get("/variables", auth, Variables);
 router.get("/settings", auth, Settings);
+
 router.get("/supporters", Supporters);
+router.post("/thank", Thanks);
 
 
 // Functions
